Update the ping url to be /solr/admin/ping
As per the solr documention, the default ping url is /solr/admin/ping.
In the event that someone wants to change the default ping path (to match
a production environment, for example), it might make sense to make this
setting configurable here as well.

Solr Documentation on the ping url:
http://wiki.apache.org/solr/SolrConfigXml#The_Admin.2BAC8-GUI_Section